test: test both vm version 0 and vm version 1 for all syscalls
