+++ script/src/syscalls/tests/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod utils;
+
+mod vm_version_0;
+#[path = "vm_latest/mod.rs"]
+mod vm_version_1;
