+++ script/src/syscalls/tests/vm_latest/mod.rs
@@ -1,0 +1,4 @@
+const SCRIPT_VERSION: crate::ScriptVersion = crate::ScriptVersion::latest();
+
+mod syscalls_1;
+mod syscalls_2;
