+++ script/src/syscalls/tests/vm_version_0.rs
@@ -1,0 +1,6 @@
+const SCRIPT_VERSION: crate::ScriptVersion = crate::ScriptVersion::V0;
+
+#[path = "vm_latest/syscalls_1.rs"]
+mod syscalls_1;
+#[path = "vm_latest/syscalls_2.rs"]
+mod syscalls_2;
